feat: build products lambdas to separate folder
