+++ README.md
@@ -1,6 +1,12 @@
 # Info
 
-Backend template for https://github.com/rolling-scopes-school/aws/tree/main/aws-developer
+Backend for AWS Developer curse: https://github.com/rolling-scopes-school/aws/tree/main/aws-developer
+
+Front-end deploy: https://d1sqwar1feok6t.cloudfront.net/
+
+## Ready services
+
+- product-service
 
 ## Full App Scheme:
