+++ product-service/tsconfig.json
@@ -6,7 +6,10 @@
       "es2020",
       "dom"
     ],
-    "declaration": true,
+    "outDir": "dist",
+    "rootDir": "./src",    
+    "declaration": false,
+    "declarationMap": false,
     "strict": true,
     "noImplicitAny": true,
     "strictNullChecks": true,
@@ -16,8 +19,8 @@
     "noUnusedParameters": false,
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": false,
-    "inlineSourceMap": true,
-    "inlineSources": true,
+    "inlineSourceMap": false,
+    "inlineSources": false,
     "experimentalDecorators": true,
     "strictPropertyInitialization": false,
     "typeRoots": [
